+++ src/GraphQL.php
@@ -3,7 +3,6 @@
 declare(strict_types = 1);
 namespace Rebing\GraphQL;
 
-use Error as PhpError;
 use Exception;
 use GraphQL\Error\DebugFlag;
 use GraphQL\Error\Error;
@@ -254,6 +253,7 @@ class GraphQL
      */
     public function getGlobalResolverMiddlewares(): array
     {
+        /** @var list<class-string|object> $resolverMiddlewares */
         $resolverMiddlewares = $this->config->get('graphql.resolver_middleware_append') ?? [];
 
         return array_merge($resolverMiddlewares, $this->globalResolverMiddlewares);
@@ -365,7 +365,7 @@ class GraphQL
                 }
 
                 if (isset($objectType->config[$key])) {
-                    $objectType->config[$key] = $value;
+                    $objectType->config[$key] = $value; // @phpstan-ignore assign.propertyType (only 'name'/'description' keys are set in practice)
                 }
             }
         } elseif (\is_array($type)) {
@@ -425,7 +425,7 @@ class GraphQL
             $typeFields[$name] = $field;
         }
 
-        return new ObjectType(array_merge([
+        return new ObjectType(array_merge([ // @phpstan-ignore argument.type (dynamic field config array can't be statically verified)
             'fields' => $typeFields,
         ], $opts));
     }
@@ -468,7 +468,7 @@ class GraphQL
             $directives[$directive->name] = $directive;
         }
 
-        return new Schema([
+        return new Schema([ // @phpstan-ignore argument.type (objectType() always returns ObjectType but is typed as Type)
             'query' => $query,
             'mutation' => $mutation,
             'subscription' => $subscription,
@@ -637,6 +637,7 @@ class GraphQL
      */
     public static function handleErrors(array $errors, callable $formatter): array
     {
+        /** @var ExceptionHandler $handler */
         $handler = app()->make(ExceptionHandler::class);
 
         foreach ($errors as $error) {
@@ -644,12 +645,7 @@ class GraphQL
             $error = $error->getPrevious() ?: $error;
 
             // Don't report certain GraphQL errors
-            if ($error instanceof ValidationError ||
-                $error instanceof AuthorizationError ||
-                !(
-                    $error instanceof Exception ||
-                    $error instanceof PhpError
-                )) {
+            if ($error instanceof ValidationError || $error instanceof AuthorizationError) {
                 continue;
             }