+++ src/Support/UnionType.php
@@ -35,6 +35,6 @@ abstract class UnionType extends Type
 
     public function toType(): GraphqlType
     {
-        return new BaseUnionType($this->toArray());
+        return new BaseUnionType($this->toArray()); // @phpstan-ignore argument.type (toArray() builds a valid config, but its dynamic shape can't be statically verified)
     }
 }